refactor: type configurations in a single folder
