+++ src/types/heroSection.types.ts
