+++ src/types/privacy.types.ts
